Running unit test
